refactor db service and add site flag start date and end date validator to api
