+++ api/package.json
@@ -17,6 +17,7 @@
     "koa-logger": "^3.1.0",
     "koa-router": "^7.2.1",
     "koa2-cors": "^2.0.3",
+    "moment": "^2.19.1",
     "mongoose": "^4.12.4",
     "ramda": "^0.25.0"
   },
