+++ api/services/index.js
@@ -1,0 +1,3 @@
+import DB from './db'
+
+export default { DB }
